+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.3
+* Tested with WordPress 5.6
+* Admin scripts and styles are launched on the plugin settings page only.
+
 ## 1.2
 * Tested with WordPress 5.5